Refactor announcement types
